It's getting bumpy in here.
So, we can make bmp's as well.  I'm not sure we're making them correctly though as the padding stuff seems to happen in the wrong place when you look at the image in Preview or whatever.  Also, the bmp you get from README doesn't exactly look right to me (one colour? seriously?).

How the hell do I test this?