And Lo! There was midriffs.
So, a pretty naive implementation of MIDI file generation makes it's way into the stegosaurus. It's far from perfect as it probably won't generate files that are listen-to-able (12 hour mostly silence epics notwithstanding).  We should probably inviestigate writing out functioning events, rather than just appending data to a header.  It's ok for WAV and (to a lesser degree) BMP, but clearly not here.

Also, now we have 3 generators it might be time to refactor the similarities and get with the programming skills.  Otherwise folk'll think I'm a copy'n'paste mong.