Slight refactoring, all the 'makers' now extend Stegosaurus::Genus and get the file name generationo from there.
I might be messing up the metaphor here, I suspect that a Stegosaurus is a species, not a family (which is what calling bumps, waves & midriffs genuses would imply.  Ah well.)

There might be more refactoring to be done.  I'm sure the simplest case of writing a file can be genus-ified.