Add the variable length quantity value
We'll need this shortly so we can write delta-time values properly.  These
tell midi when to trigger the event it's attached to.  These are written
as variable length quantity values meaning they can take up more than one
byte, but to do so we set the first bit to 1 for all but the last byte
and the numeric value is contained in the remaining 7bits of the byte. We
had a method for doing this, but it was slightly wrong and wouldn't
generate the value correctly.  E.g. trying to store 0x80 (which needs 8
bits to store) would store:

    0x80 0x01

Instead of:

    0x81 0x0

Which is what it should do.  The new version gets it correct - tbh I took
my fix from the implementation in https://github.com/jimm/midilib for when
it writes variable length quantity values.

https://github.com/jimm/midilib/blob/6c8e481ae72cd9f00a38eb3700ddfca6b549f153/lib/midilib/utils.rb#L18-L33