Generate midi events from data properly
We knew we couldn't just drop the data from the file onto the end of the
header and expect it to work.  We were filtering out bytes that might look
like midi sysevents.  However, the structure of midi events is more
complex than that and instead of filtering out bytes we can be smarter.

A track is made up of the header, the length, and a series of midi events.
The midi events we care about are the ones that'll actually make some
simple music.  They're made up of a delta-time, an optional status, then
the relevant data bytes for that status.  Status bytes all have their
first bit set to 1 (e.g. as a byte they're > 128).  Data bytes all have
their first bit set to 0 (e.g. as a byte they're < 128).  We obviously
can't expect the raw data of a file to look like this, so we need to do
something.  Turns out a valid midi event that will either turn on a note
or turn one off at some moment in time will look like this:

    delta-time-for(8bits), 100 + 5bits, 0 + 7bits, 0 + 7bits

This means we can safely turn every 27 bits of the raw data into a single
midi event.  Obvs, 27 bits is 3 and 3/8th bytes which isn't great, and
unfortunately, the lowest common multiple of 27 and 8 is 216 which is 27
bytes.  This means we can read 27 bytes of data and turn that into 8 midi
events.  If the data doesn't happen to have 27 bytes we can just pad it
with 0x0 much like we do with bumps.

Knowing that we want things in chunks of 27 bytes it makes sense to change
the read bufer size to a multiple of 27 bytes, and 216 seems reasonable.