Set length in track header correctly
We can't just set it to the size of the data file, because we're going to
make it bigger.

1. We turn 1 byte into 1 or 2 so they can be stored as variable length
   quantities to represent delta times
2. For every 27bits of data, we add 5bits to store it as a midi event.

We could calculate the extra 5bits based on the size of the existing file,
but we can't know if the delta-times will need an extra byte without
reading the data file.  If we have to read it, we might as well create the
track event data while we do it so we write it to a tempfile, then
we can generate the header based on the size of that tempfile and write it
to the new data file.  Finally we can write the data from the tempfile
onto the end of the new data file.

Lucky that we extracted all those methods from `genus` earlier to make
this easy, eh?