Only write padding bytes if we need to
We don't need to complete the last pixel if the data is lucky enough to be
a complete number of pixels.  Writing 0 bytes would do the same thing but
we can avoid it entirely.  Similarly, we don't need to add the bytes for
the final padding pixels if the data is such a size as to not need any.