Add convenience functions for opening files with some apps
This is to save me from having to come out of the shell to open the file.

We started with using quicklook via `qlmanage -p` on the supplied path name.  Which is fine, but quicklook windows will open on the screen I'm using, and that's annoying when presenting.

Now I have `open_xxx_file` methods that use `open -a <some app> <path>` and I've optimised for apps that will stay sticky to a given desktop when you open a file with them.  Not perfect, but will do for now.