Make writing end of track marker optional (default false)
Turns out, MIDITrail doesn't like our files if they have EOT markers, but
is happy with them otherwise. Who knows why, but there we are. For now
make it optional, default to false. We _might_ debug this later as it
suggests a deeper malaise with the file. Probably the "Too shorten midi
file" error we get from timidity has _something_ to do with it.