Calculate image size correctly
line_pad_bytes might be 0 when in fact we need to add less than a byte to
get to a 4byte scale line. Instead of using that we'll instead calculate
it based on width * bit_count which is bit_scale and then round that to
the closest 32-bit (4-byte) number.