Get rid of "image_details" tuple
Mostly, we didn't need all the things being passed around into these
methods so lets just pass what we do need. It may have been useful for
debugging output in the past (and may be again), but it's noise in the
code.

We also take the opportunity to make scan_line_pad bit scale while we're
doing this. if bit count is < 8 then we might need less than a byte to
complete a scanline so lets keep this bit scale.

The rest of the code needs tidying up to properly support less than 8-bit
colour depths, but this is the first step.