Support bit-scale and byte-scale pixel data properly
We take some tricks from midriffs for dealing with bit scale colour
depths and convert bytes into bits. Because we'll actually be potentially
writing less than the bytes we extracted from the file at this scale, we
need to keep an array of bits around to fill and empty as we write rows.
This fill and empty happens at a slightly different cadence to reading
chunks of bytes. It's annoying, but works.

We also, finally, do some cleanup of the while loops and "write the last
data" code so it's not duplicated everywhere.