Correct handling of bit-scale scan line padding
We were calculating it as `(32 - (bits_needed % 32))` which, if
bits_needed was already a multiple of 32 would mean adding 4 whole extra
bytes of line padding we didn't need. `scan_line_pad_bits` already
handles this edgecase but because it took width and we didn't have width
in `write_bit_scale_bump_data` we tried to calculate it again.

The solution is to change the method signature to take width instead of
bits_needed.

This then makes `write_bit_scale_bump_data` and
`write_byte_scale_bump_data` more similar in method signatures, so we
lean more into that by re-ordering the args and drop passing
`line_pad_bits` around and just calculate when needed in the two methods.
We were calculating it early, and passing it down, but not actually using
it in both methods. We _do_ now use it in both methods, but we can
calculate as needed.