Fix bit-scale scanlines for big files
As we build up `row` with the "left over" bits from the difference between how much data we can fetch in one go and how much we use to write a scan line, it's possible, for a large file, that we end up with more bits in `row` than we need for a whole scanline.  Once we hit that threshold we can just write a complete line without looking at fetching more from the source file.

This spotted by me running the 1-bit version of the bumps generator across the ruby interpreter while presenting this code at RubyConf.  That was a mistake but it "usefully" exposed this bug, so that's nice.